fix dev names to be final ones
